+++ app/ui/board/Search/index.tsx
@@ -3,10 +3,10 @@ import { useState, useEffect } from 'react';
 import { Button } from '@/app/ui/components/Button';
 import DropDown from '@/app/ui/components/DropDown';
 import { MenuItem } from '@headlessui/react';
-import UpdateBoardModal from '@/app/ui/components/Modal/update-board';
+import UpdateBoardModal from '@/app/ui/board/Update';
 import { useSharedState } from '@/app/ui/components/SharedState/Context';
 import { publish } from '@/app/lib/data/board/main';
-import Share from '@/app/ui/components/Modal/share-with-user';
+import Share from '@/app/ui/board/Share';
 import Link from 'next/link';
 import DeleteBoard from '@/app/ui/board/Delete';
 import RequestBoardContribution from '@/app/ui/board/Request';
